layout : 커뮤니티 페이지 페이지 레이아웃 다 만들어 놓음
