+++ app/src/main/java/com/ing/offroader/ui/activity/chatbot/viewmodel/ChatBotViewModel.kt
@@ -1,4 +1,4 @@
-package com.ing.offroader.ui.fragment.chatbot.viewmodel
+package com.ing.offroader.ui.activity.chatbot.viewmodel
 
 import android.util.Log
 import androidx.lifecycle.LiveData
