+++ app/src/main/java/com/ing/offroader/ui/fragment/chatbot/adapter/CommunityAdapter.kt
@@ -1,0 +1,6 @@
+package com.ing.offroader.ui.fragment.chatbot.adapter
+
+import com.ing.offroader.ui.fragment.chatbot.viewmodel.CommunityViewModel
+
+class CommunityAdapter(viewModel: CommunityViewModel) {
+}
