+++ app/src/main/java/com/ing/offroader/ui/fragment/chatbot/viewmodel/CommunityViewModel.kt
@@ -1,0 +1,4 @@
+package com.ing.offroader.ui.fragment.chatbot.viewmodel
+
+class CommunityViewModel {
+}
